Cut out all Firestore access functions into Firestore.ts
