+++ src/features/session/SessionRepository.ts
@@ -22,25 +22,10 @@
 // THE SOFTWARE.
 //
 
-import { map, Observable } from "rxjs";
-import { Conference, conferenceConverter } from "../../entities/Conference";
-import { Session, sessionConverter } from "../../entities/Session";
-import { Event, eventConverter } from "../../entities/Event";
-import { getFirestore, withFirestore } from "../../Firebase";
-import { collection, getDocs, limit, onSnapshot, orderBy, query, startAfter, where } from "firebase/firestore";
-import { Query, QuerySnapshot } from "@firebase/firestore";
-import { CaseInsensitiveSearch } from "../../utils/CaseInsensitiveSearch";
-
-export interface SessionFilter {
-  conferenceId?: string,
-  minutes?: number,
-  keywords?: string[],
-}
-
-export interface FilteredSessions {
-  sessions: Session[];
-  more: (() => Promise<FilteredSessions>) | undefined;
-}
+import { Observable } from "rxjs";
+import { Conference } from "../../entities/Conference";
+import { Event } from "../../entities/Event";
+import { FilteredSessions, Firestore, SessionFilter } from "../../Firestore";
 
 export interface SessionRepository {
   getAllConferences$(): Observable<Conference[]>;
@@ -50,98 +35,14 @@ export interface SessionRepository {
 
 export class FirestoreSessionRepository implements SessionRepository {
   getAllConferences$(): Observable<Conference[]> {
-    return withFirestore(firestore => {
-      const collectionRef = collection(firestore, "conferences");
-      const docQuery = query(collectionRef,
-        orderBy("starts", "desc"),
-      ).withConverter(conferenceConverter);
-      return new Observable<QuerySnapshot<Conference>>(subscriber => {
-        return onSnapshot(docQuery, subscriber);
-      }).pipe(
-        map(snapshot => {
-          return snapshot
-            .docs
-            .map(it => it.data())
-        }),
-      );
-    });
+    return Firestore.getAllConferences$();
   }
 
-  async getAllEvents(): Promise<Event[]> {
-    const firestore = await getFirestore();
-    const collectionRef = collection(firestore, "events");
-    const docQuery = query(collectionRef,
-      where("hidden", "==", false),
-      orderBy("starts", "desc"),
-    ).withConverter(eventConverter);
-    const snapshot = await getDocs(docQuery);
-    return snapshot.docs.map(doc => doc.data());
+  getAllEvents(): Promise<Event[]> {
+    return Firestore.getAllEvents();
   }
 
-  async getSessions(filter: SessionFilter): Promise<FilteredSessions> {
-    const LIMIT = 100;
-    const firestore = await getFirestore();
-    const collectionRef = collection(firestore, "sessions");
-    let docQuery: Query<Session> = collectionRef.withConverter(sessionConverter);
-    if (filter.conferenceId !== undefined) {
-      docQuery = query(docQuery, where("conferenceId", "==", filter.conferenceId));
-    }
-    if (filter.minutes !== undefined) {
-      docQuery = query(docQuery, where("minutes", "==", filter.minutes));
-    }
-    docQuery = query(docQuery,
-      orderBy("starts", "asc"),
-      limit(LIMIT),
-    );
-    return await filterSessions(docQuery);
-    
-    async function filterSessions(docQuery: Query<Session>): Promise<FilteredSessions> {
-      const snapshot = await getDocs(docQuery);
-      const sessions = snapshot
-        .docs
-        .map(doc => doc.data())
-        .filter(filterByKeywords(filter.keywords));
-      let more: (() => Promise<FilteredSessions>) | undefined = undefined;
-      if (snapshot.size === LIMIT) {
-        const nextQuery = query(docQuery, startAfter(snapshot.docs[snapshot.docs.length - 1]));
-        more = async () => await filterSessions(nextQuery);
-      }
-      return { sessions, more };
-    }
-
-    function filterByKeywords(keywords?: string[]): (session: Session) => boolean {
-      if (keywords === undefined) {
-        return _ => true;
-      }
-
-      return session => {
-        for (const keyword of keywords) {
-          const searcher = new CaseInsensitiveSearch(keyword);
-
-          if (!searcher.foundIn(session.title)) {
-            if (!searcher.foundIn(session.description)) {
-              let found = false;
-              for (const speaker of session.speakers) {
-                if (searcher.foundIn(speaker.name)) {
-                  found = true;
-                  break;
-                }
-                if (speaker.twitter !== undefined) {
-                  if (searcher.foundIn(speaker.twitter)) {
-                    found = true;
-                    break;
-                  }
-                }
-              }
-              if (!found) {
-                return false;
-              }
-            }
-          }
-        }
-
-        return true;
-      };
-    }
+  getSessions(filter: SessionFilter): Promise<FilteredSessions> {
+    return Firestore.getSessions(filter);
   }
 }
